% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS.R
\name{CellSIUS}
\alias{CellSIUS}
\title{CellSIUS: Cell Subtype Identification from Upregulated gene Sets}
\usage{
CellSIUS(mat.norm, group_id, min_n_cells = 10, min_fc = 2,
  corr_cutoff = NULL, iter = 0, max_perc_cells = 50,
  fc_between_cutoff = 1, mcl_path = "~/local/bin/mcl")
}
\arguments{
\item{mat.norm}{Numeric Matrix: normalized gene expression matrix where \code{rownames} and \code{colnames} are gene and cell ID, respectively.}

\item{group_id}{Character: vector with cluster cell assignment. Make sure that the order of cell cluster assignment reflects the order of the columns of norm.mat.}

\item{min_n_cells}{Integer: when identifying bimodal gene distributions, this specifies the minimum number of cells per mode. Clusters with a total number of cells below \code{min_n_cells} will be entirely ignored. Defaults to 10.}

\item{min_fc}{Numeric: minimum difference in mean [log2] between the two modes of the gene expression distribution. Defaults to 2.}

\item{corr_cutoff}{Numeric: correlation cutoff for MCL clustering of candidate marker genes. If \code{NULL}, it will be set automatically for each cluster, which in general works much better than forcing a fixed value. Therefore, leave this at the default unless having a good reason to change it. Defaults to \code{NULL}.}

\item{iter}{0 or 1: relevant for the final step (assigning cells to subgroups). If set to 1, the first mode of gene expression will be discarded. Cells are then assigned based on the 2nd and 3rd mode, which results in a more stringent assignment. The default is 0 and is usually stringent enough. Defaults to 0.}

\item{max_perc_cells}{Numeric: maximum percentage of cells that can be part of a subcluster. Defaults to 50, implying that a “subgroup” cannot contain more than half of the total observations.}

\item{fc_between_cutoff}{Numeric: minimum difference [log2] in gene expression between cells in the subcluster and all other cells. The higher, the more cluster-specific is the gene signature. Note that this should not be set higher than min_fc.}

\item{mcl_path}{Character: path to the MCL executable. The external GNU MCL software for UNIX needs to be installed by the user.}
}
\description{
\strong{Identification of cluster-specific gene sets and cell classification according to their expression.}

CellSIUS enables the identification and characterization
of (rare) cell sub-populations from complex scRNA-seq datasets: it takes as input expression values of N cells grouped into M(>1) clusters.
Within each cluster, genes with a bimodal distribution are selected and only genes with cluster-specific
expression are retained. Among these candidate marker genes, sets with correlated expression patterns
are identified by graph-based clustering. Finally, cells are assigned to subgroups based on their average
expression of each gene set. The CellSIUS algorithm output provides the rare/ sub cell types by cell indices and
their transcriptomic signatures.
}
\details{
\code{CellSIUS} returns a \code{data.table} of a collection of key/ value pairs:
\itemize{
  \item \strong{cell_idx} Cell indices correspond to the \code{colnames} of \code{norm.mat}
  \item \strong{gene_id} Gene IDs correspond to the \code{rownames} of \code{norm.mat}
  \item \strong{main_cluster} Name of the main cluster defined in \code{group_id} input
  \item \strong{expr} Average gene expression of ‘gene_id’ across the cells in the correspondent ‘main_cluster’
  \item \strong{sub_cluster} identifies if the cell is \emph{not} a member of a CellSIUS subcluster. Subclusters are named as follows: [Name-of-main-cluster]_[\emph{number of subcluster}]_[0 or 1], where in the last position 0 means the cell is not a member, 1 means it is a member.
  \item \strong{N_cells} Number of cells in the correspondent CellSIUS sub cluster
  \item\strong{log2FC}   log2 fold change of between the two modes of the bimodal gene expression distribution across cells in the correspondent ‘main_cluster’
}
}
\section{Warning}{

The execution of CellSIUS requires
the external GNU Markov Cluster Algorithm (MCL) software for UNIX. It needs to be installed by the user [\href{https://micans.org/mcl/}{link}].
}

\examples{
require(CellSIUS)
require(ggplot2)
require(data.table)

# 2D cell projection
qplot(my_tsne[,1],my_tsne[,2],xlab="tSNE1",ylab="tSNE1",main='2D scRNAseq cell projection')
qplot(my_tsne[,1],my_tsne[,2],xlab="tSNE1",ylab="tSNE1",color=as.factor(clusters),main="Cells colored by clusters")

# Run CellSIUS
# !WARNING: before to run the CellSIUS function, take care to correctly set the mcl_path to point
# to the executable of your local installation of MCL algorithm

CellSIUS.out<-CellSIUS(mat.norm = norm.counts,group_id = clusters,min_n_cells=10, min_fc = 2,
             corr_cutoff = NULL, iter=0, max_perc_cells = 50,
             fc_between_cutoff = 1,mcl_path = "~/local/bin/mcl")

#____________________________
# EXPLORE CellSIUS RESULTS:
#____________________________

# Summary of sub-populations identified by CellSIUS

Result_List=CellSIUS_GetResults(CellSIUS.out=CellSIUS.out)

# 2D visualization of of sub-populations identified by CellSIUS
require(RColorBrewer)
CellSIUS_plot(coord = my_tsne,CellSIUS.out = CellSIUS.out)

# Final clustering assignment

Final_Clusters = CellSIUS_final_cluster_assignment(CellSIUS.out=CellSIUS.out, group_id=clusters, min_n_genes = 3)
table(Final_Clusters)
qplot(my_tsne[,1],my_tsne[,2],xlab="tSNE1",ylab="tSNE1",color=as.factor(Final_Clusters))

}
\references{
Wegmann, R.\emph{et Al.}, 2018.
CellSIUS for sensitive and specific identification and characterization
of rare cell populations from complex single-cell RNA sequencing data.
\emph{Nature Comunications Submission}
}
\seealso{
Accessory functions are provided to help the user to summarize and visualize the CellSIUS results:
\itemize{
  \item \code{\link[CellSIUS]{CellSIUS_GetResults}}
  \item \code{\link[CellSIUS]{CellSIUS_plot}}
  \item \code{\link[CellSIUS]{CellSIUS_final_cluster_assignment}}
}
}
