% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS_final_cluster_assignment.R
\name{CellSIUS_final_cluster_assignment}
\alias{CellSIUS_final_cluster_assignment}
\title{Final assignment to unique clusters}
\usage{
CellSIUS_final_cluster_assignment(CellSIUS.out, group_id, min_n_genes = 3)
}
\arguments{
\item{CellSIUS.out}{data.table: Output of \code{\link[CellSIUS]{CellSIUS}}.}

\item{group_id}{Character: vector with cluster cell assignment.
Make sure that the order of cell cluster
assignment reflects the order of the columns of norm.mat.}

\item{min_n_genes}{Integer: Minimum number of genes in a signature.}
}
\description{
\code{CellSIUS_final_cluster_assignment} merges the main clusters with the sub clusters
identified by \code{\link[CellSIUS]{CellSIUS}} and returns the final cluster assignment.
}
\details{
\code{CellSIUS_final_cluster_assignment} returns a \code{character}.
Note that a cell can potentially be a member of multiple subgroups.
}
\seealso{
Main CellSIUS function: \code{\link[CellSIUS]{CellSIUS}}.
}
