% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS.R
\name{cellsius_find_bimodal_genes}
\alias{cellsius_find_bimodal_genes}
\title{Identify genes with bimodal distribution}
\usage{
cellsius_find_bimodal_genes(expr, min_n_cells, max_perc_cells)
}
\description{
Accessory function of \code{\link[CellSIUS]{CellSIUS}}
}
\keyword{internal}
