% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS.R
\name{cellsius_find_gene_sets}
\alias{cellsius_find_gene_sets}
\title{MCL clustering to find correlated gene sets}
\usage{
cellsius_find_gene_sets(expr_dt_melt, corr_cutoff = NULL, min_corr = 0.35,
  max_corr = 0.5, mcl_path = mcl_path)
}
\description{
Accessory function of \code{\link[CellSIUS]{CellSIUS}}
}
\keyword{internal}
