% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS_GetResults.R
\name{cellsius_get_cells_per_subclust}
\alias{cellsius_get_cells_per_subclust}
\title{Get cells per subcluster}
\usage{
cellsius_get_cells_per_subclust(CellSIUS.out)
}
\description{
Accessory function of \code{\link[CellSIUS]{CellSIUS_GetResults}}
}
\keyword{internal}
