% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS_GetResults.R
\name{CellSIUS_GetResults}
\alias{CellSIUS_GetResults}
\title{Summary of CellSIUS results}
\usage{
CellSIUS_GetResults(CellSIUS.out)
}
\arguments{
\item{CellSIUS.out}{data.table: Output of \code{\link[CellSIUS]{CellSIUS}}}
}
\description{
\code{CellSIUS_GetResults} function prints a summary of the CellSIUS results.
It returns a nested list where each entry of this list corresponds to one subcluster.
For each subcluster, the cell index of its members and the ID and log2FC of its marker
genes are returned.
}
\seealso{
Main CellSIUS function: \code{\link[CellSIUS]{CellSIUS}}
}
