% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS_plot.R
\name{CellSIUS_plot}
\alias{CellSIUS_plot}
\title{2D visualization of CellSIUS sub clusters}
\usage{
CellSIUS_plot(coord, CellSIUS.out)
}
\arguments{
\item{coord}{matrix: 2D cell distribution coordinates, \emph{e.g.} tSNE or PCA,
where its \code{rownames} correspond to cell indices.}

\item{CellSIUS.out}{data.table: Output of \code{\link[CellSIUS]{CellSIUS}}}
}
\description{
\code{CellSIUS_plot} function returns a ggplot visualization of the 2D distribution
of cells highlighting the sub clusters identified by CellSIUS.
}
\seealso{
Main CellSIUS function: \code{\link[CellSIUS]{CellSIUS}}
}
