% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS_GetResults.R
\name{cellsius_print_summary}
\alias{cellsius_print_summary}
\title{Print Summary}
\usage{
cellsius_print_summary(CellSIUS.out)
}
\description{
Accessory function of \code{\link[CellSIUS]{CellSIUS_GetResults}}
}
\keyword{internal}
