% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS.R
\name{cellsius_sub_cluster}
\alias{cellsius_sub_cluster}
\title{Assign cells to subgroups}
\usage{
cellsius_sub_cluster(mean_expr, sub_cluster, gene_cluster, iter = iter)
}
\description{
Accessory function of \code{\link[CellSIUS]{CellSIUS}}
}
\keyword{internal}
