% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSIUS.R
\name{cellsius_test_cluster_specificity}
\alias{cellsius_test_cluster_specificity}
\title{Check whether the subset of genes are specific to one cell subgroup}
\usage{
cellsius_test_cluster_specificity(exprs, cluster, current_cluster,
  fc_between_cutoff)
}
\description{
Accessory function of \code{\link[CellSIUS]{CellSIUS}}
}
\keyword{internal}
